[frontend] Fix tensor list alias annotation
The current logic checks alias info before `[]` and after. If no alias info exists after `[]`, we overwrite the alias info before. This logic failed on argument like `Tensor(a!)[]`, dropping the alias info before `[]` on the floor. This PR adds a new alias info if it's missing after `[]`. This way we can keep the alias info before `[]`.

ghstack-source-id: 5f6ecdf9f0c4f2a8c2ef0964ae67cf630f8c363c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83921